🔨 Move compat.contextlib from retrocookie.pr to retrocookie
