+++ src/retrocookie/pr/compat/contextlib.py
@@ -1,18 +1,4 @@
 """Compatibility shim for contextlib."""
-import contextlib
-from typing import Callable
-from typing import ContextManager
-from typing import Iterator
-from typing import TypeVar
+from retrocookie.compat.contextlib import contextmanager
 
-
-T = TypeVar("T")
-
-
-def contextmanager(
-    func: Callable[..., Iterator[T]]
-) -> Callable[..., ContextManager[T]]:
-    """Fix annotations of functions decorated by contextlib.contextmanager."""
-    result = contextlib.contextmanager(func)
-    result.__annotations__ = {**func.__annotations__, "return": ContextManager[T]}
-    return result
+__all__ = ["contextmanager"]
